create and test setMaxInitialSupply and maxInitialSupply methods.

+++ contract/contracts/ERC1155MintRange.sol
@@ -16,32 +16,32 @@ abstract contract ERC1155MintRange is ERC1155_ {
         "No initial holders set. Use _setInitialHolders()";
 
     // Mapping from token ID to balancesInitialzed flag
-    mapping(uint256 => mapping(address => bool)) public _balancesInitialized;
+    mapping(uint => mapping(address => bool)) public _balancesInitialized;
 
     // Mapping from token ID to totalSupply
-    mapping(uint256 => int256) public _totalSupply;
+    mapping(uint => int256) public _totalSupply;
 
     // Mapping to keep track of tokens that are minted via ERC1155._mint() or  ERC1155._mintBatch()
-    mapping(uint256 => bool) public _manualMint;
-    uint256 public _manualMints;
+    mapping(uint => bool) public _manualMint;
+    uint public _manualMints;
 
     // Track initial holders across tokenID ranges + lookup mapping;
     address[][] public _initialHolders;
-    uint256[] public _initialHoldersRange;
+    uint[] public _initialHoldersRange;
     mapping(uint => mapping(address => bool)) public _initialHoldersMappings;
 
-    uint256 public _lastRangeTokenId;
+    uint public _lastRangeTokenId;
     bool public _zeroMinted;
 
     /**
      * @dev Implement: Return initial token balance for address.
      * This function MUST be pure: Always return the same values for a given input.
      */
-    function initialBalanceOf(address account, uint256 tokenId)
+    function initialBalanceOf(address account, uint tokenId)
         public
         view
         virtual
-        returns (uint256);
+        returns (uint);
 
     /**
      * @dev Set initial holders. mintRange will distribute tokens to these holders
@@ -49,7 +49,7 @@ abstract contract ERC1155MintRange is ERC1155_ {
     function _setInitialHolders(address[] memory addresses) internal virtual {
         _initialHoldersRange.push(_zeroMinted ? _lastRangeTokenId + 1 : 0);
         _initialHolders.push(addresses);
-        for(uint256 i=0; i<addresses.length; i++) {
+        for(uint i=0; i<addresses.length; i++) {
             _initialHoldersMappings[_initialHolders.length-1][addresses[i]] = true;
         }
     }
@@ -58,8 +58,8 @@ abstract contract ERC1155MintRange is ERC1155_ {
      * @dev Lazy-mint a range of new tokenIds to initial holders
      */
     function _mintRange(
-        uint256[] memory ids,
-        uint256[][] memory amounts,
+        uint[] memory ids,
+        uint[][] memory amounts,
         bytes32 inputChecksum
     ) internal virtual {
         address[] memory addresses = initialHolders();
@@ -81,7 +81,7 @@ abstract contract ERC1155MintRange is ERC1155_ {
         if (_zeroMinted == false) {
             _zeroMinted = true;
         }
-        for (uint256 i = 0; i < addresses.length; i++) {
+        for (uint i = 0; i < addresses.length; i++) {
             emit TransferBatch(
                 msg.sender,
                 address(0),
@@ -95,19 +95,19 @@ abstract contract ERC1155MintRange is ERC1155_ {
     /**
      * @dev Returns true if tokenId was minted.
      */
-    function exists(uint256 tokenId) public view virtual returns (bool) {
+    function exists(uint tokenId) public view virtual returns (bool) {
         return _inRange(tokenId) || _manualMint[tokenId] == true;
     }
 
     /**
      * @dev See {ERC1155-balanceOf}.
      */
-    function balanceOf(address account, uint256 id)
+    function balanceOf(address account, uint id)
         public
         view
         virtual
         override
-        returns (uint256)
+        returns (uint)
     {
         require(
             account != address(0),
@@ -129,7 +129,7 @@ abstract contract ERC1155MintRange is ERC1155_ {
     /**
      * @dev Returns initial holders of a token.
      */
-    function initialHolders(uint256 tokenId)
+    function initialHolders(uint tokenId)
         public
         view
         virtual
@@ -151,7 +151,7 @@ abstract contract ERC1155MintRange is ERC1155_ {
     /**
      * @dev Returns true if address is an initial holder of tokenId
      */
-    function isInitialHolderOf(address account, uint256 tokenId)
+    function isInitialHolderOf(address account, uint tokenId)
         public
         view
         returns (bool)
@@ -164,56 +164,56 @@ abstract contract ERC1155MintRange is ERC1155_ {
     /**
      * @dev Total amount of tokens with a given id.
      */
-    function totalSupply(uint256 tokenId)
+    function totalSupply(uint tokenId)
         public
         view
         virtual
-        returns (uint256)
+        returns (uint)
     {
         // Pre initialization
         if (_inRange(tokenId) && !_manualMint[tokenId]) {
-            uint256 initialTotalSupplySum = 0;
+            uint initialTotalSupplySum = 0;
             address[] memory initialHolderAddresses = initialHolders(tokenId);
-            for (uint256 i = 0; i < initialHolderAddresses.length; i++) {
+            for (uint i = 0; i < initialHolderAddresses.length; i++) {
                 initialTotalSupplySum += initialBalanceOf(
                     initialHolderAddresses[i],
                     tokenId
                 );
             }
-            return uint256(int256(initialTotalSupplySum) + _totalSupply[tokenId]);
+            return uint(int256(initialTotalSupplySum) + _totalSupply[tokenId]);
         }
 
         // manually minted
-        return uint256(_totalSupply[tokenId]);
+        return uint(_totalSupply[tokenId]);
     }
 
     /**
      * @dev Generate mintRange inputs for x new tokens.
      */
-    function getMintRangeInput(uint256 numberOfTokens)
+    function getMintRangeInput(uint numberOfTokens)
         public
         view
         returns (
-            uint256[] memory,
-            uint256[][] memory,
+            uint[] memory,
+            uint[][] memory,
             bytes32
         )
     {
-        uint256 firstId = _zeroMinted ? _lastRangeTokenId + 1 : 0;
+        uint firstId = _zeroMinted ? _lastRangeTokenId + 1 : 0;
         address[] memory holders = initialHolders();
-        uint256[] memory ids = new uint256[](numberOfTokens);
-        uint256[][] memory amounts = new uint256[][](holders.length);
+        uint[] memory ids = new uint[](numberOfTokens);
+        uint[][] memory amounts = new uint[][](holders.length);
 
-        uint256 newIndex = 0;
-        for (uint256 i = 0; newIndex < numberOfTokens; i++) {
-            uint256 newId = firstId + i;
+        uint newIndex = 0;
+        for (uint i = 0; newIndex < numberOfTokens; i++) {
+            uint newId = firstId + i;
             if (_manualMint[newId]) {
                 continue;
             }
             ids[newIndex] = newId;
-            for (uint256 b = 0; b < holders.length; b++) {
+            for (uint b = 0; b < holders.length; b++) {
                 if (newIndex == 0) {
-                    amounts[b] = new uint256[](numberOfTokens);
+                    amounts[b] = new uint[](numberOfTokens);
                 }
                 amounts[b][newIndex] = initialBalanceOf(holders[b], newId);
             }
@@ -241,12 +241,12 @@ abstract contract ERC1155MintRange is ERC1155_ {
         address operator,
         address from,
         address to,
-        uint256[] memory ids,
-        uint256[] memory amounts,
+        uint[] memory ids,
+        uint[] memory amounts,
         bytes memory data
     ) internal virtual override {
-        for (uint256 i = 0; i < ids.length; ++i) {
-            uint256 id = ids[i];
+        for (uint i = 0; i < ids.length; ++i) {
+            uint id = ids[i];
 
             // when minting
             if (from == address(0)) {
@@ -272,7 +272,7 @@ abstract contract ERC1155MintRange is ERC1155_ {
     /**
      * @dev Writes dynamic initial Balance to state if uninitialized.
      */
-    function _maybeInitializeBalance(address account, uint256 id) private {
+    function _maybeInitializeBalance(address account, uint id) private {
         if (
             account != address(0) &&
             _inRange(id) &&
@@ -280,7 +280,7 @@ abstract contract ERC1155MintRange is ERC1155_ {
             !_manualMint[id] &&
             isInitialHolderOf(account, id)
         ) {
-            uint256 balance = initialBalanceOf(account, id);
+            uint balance = initialBalanceOf(account, id);
             if (balance > 0) {
                 _balancesInitialized[id][account] = true;
                 _balances[id][account] = balance;
@@ -291,19 +291,19 @@ abstract contract ERC1155MintRange is ERC1155_ {
     /**
      * @dev Returns true if token is in existing id range.
      */
-    function _inRange(uint256 tokenId) private view returns (bool) {
+    function _inRange(uint tokenId) private view returns (bool) {
         return _zeroMinted && tokenId <= _lastRangeTokenId;
     }
 
     /**
      * @dev Utility find range/bucket for tokenId.
      */
-    function _findInRange(uint256[] memory range, uint256 tokenId)
+    function _findInRange(uint[] memory range, uint tokenId)
         internal
         pure
-        returns (uint256)
+        returns (uint)
     {
-        for (uint256 i = range.length - 1; i >= 0; i--) {
+        for (uint i = range.length - 1; i >= 0; i--) {
             if (tokenId >= range[i]) {
                 return i;
             }
